Update day 04 part 1
This commit updates day 04 part 1 to make the implementation of part 2 a
little bit easier. Also, instead of converting all the numbers to
integers, this just leaves them in the string form and uses re to parse
through things (I just thought that was cool - not sure if it's better
though).