Add an init script
This commit adds a convenience script that initializes the files needed
to solve the next puzzle. It copies the template over, extracts the name
of the puzzle from the advent of code website and adds it to the code
file, downloads the sample input from the website, and downloads the
puzzle input (if it has access to a session cookies file).