Update day 12 part 1
This commit adds another approach to the problem. It is a string based
recursive method (and was actually my first attempt at solving the
problem). I scrapped it and wrote the binary version because I wasn't
getting the correct answer. Turns out that I had made a mistake
elsewhere (note to self: if you choose a dict to store data, you are
assuming that the keys will be unique). Anyway, this commit just adds my
first recursive approach.